==> analysis/02_figures.Rmd <==
---
title: "Prototype Figures"
output: 
  html_notebook:
    theme: flatly
    toc: true
    toc_float: true
    author: "David McGaughey"
    date: "2023-10-30"
---


# Load Data

```{r}
library(tidyverse)
library(Seurat)
library(ggrepel)
library(org.Mm.eg.db)
library(ComplexHeatmap)
load('../data/00_process_output.freeze_2023_10_30.Rdata')
seurat.markersCC <- seurat.markersCC %>% bind_rows(.id = 'cluster') %>% as_tibble(rownames = 'Gene') %>% mutate(Gene = gsub('\\.\\.\\.\\d+','',Gene))
```

# Hand label each cluster (guessing the cell type)
```{r}
ct_cluster_CC <- rbind(c(0, "RPC"),
                       c(1, 'RPC'),
                       c(2, 'RPE Precursor'),
                       c(3, 'Fibroblast'),
                       c(4, 'Fibroblast'),
                       c(5, 'Optic Fissure'),
                       c(6, 'Proliferating RPC'),
                       c(7, 'Epithelial'),
                       c(8, 'Red Blood Cell'),
                       c(9, 'Corneal Progenitor'),
                       c(10, 'Blood Vessel'),
                       c(11, "PBMC"),
                       c(12, "NK/T"),
                       c(13, "Lens"
                       )) %>% data.frame()
colnames(ct_cluster_CC) <- c('seurat_clusters','CellType')
seuratCC@meta.data$CellType <- seuratCC@meta.data %>% dplyr::select(seurat_clusters) %>% left_join(ct_cluster_CC) %>% pull(CellType)

# ct_cluster <- c(0, "RPC",
#   1, 'Fibroblasts',
#   2, 'RPC',
#   3, 'RPE Precursor',
#   4, 'Cornea - JCT - TM?',
#   5, 'Optic Fissure',
#   6, 'Proliferating RPC',
#   7, 'Epithelial',
#   8, 'Red Blood Cell',
#   9, '?',
#   10, 'Vessel',
#   11, "Macrophage/Microglia")

```

# UMAP

```{r, fig.width=5, fig.height=4}

umap_plot <- Seurat::FetchData(seuratCC, vars = c("UMAP_1","UMAP_2", "CellType")) %>%
  ggplot(aes(x=UMAP_1,
             y=UMAP_2, 
             color = CellType, 
             label = CellType)) +
  cowplot::theme_cowplot() +
  scale_color_manual(values = pals::glasbey() %>% unname()) +
  scale_fill_manual(values = pals::glasbey() %>% unname() ) + theme(legend.position = "none") +
  geom_point(size = 0, alpha = 0.8) +
  geom_text_repel(data = . %>% group_by(CellType) %>%
                    summarise(UMAP_1 = mean(UMAP_1),
                              UMAP_2 = mean(UMAP_2)),
                  color = 'gray20',
                  bg.color = 'gray90') +
  xlab("UMAP 1") + ylab("UMAP 2")
umap_plot
```

## Barplot of Cell Counts
```{r}
barplot <- Seurat::FetchData(seuratCC, vars = c("UMAP_1","UMAP_2", "CellType")) %>%
  group_by(CellType) %>% 
  summarise(Count = n()) %>% 
  ggplot(aes(x=reorder(CellType, Count), y=Count)) + 
  geom_bar(stat='identity') +
  xlab("CellType") +
  coord_flip() + 
  cowplot::theme_cowplot()
barplot
```

```{r, fig.height=8, fig.width=6}
cowplot::plot_grid(umap_plot, barplot, nrow = 2, align = 'hv', rel_heights = c(1,0.5))
tiff(filename = 'umap_counts.tiff', width = 3500, height = 4000, res = 600)
print(cowplot::plot_grid(umap_plot, barplot, nrow = 2, align = 'hv', rel_heights = c(1,0.6)))
dev.off()
```

## Markers

```{r, fig.width=16, fig.height=7}
genes <- c("Pax2", "Nr2f1", "Zic1", "Cp", 
           "Vim", "Ptn", "Minar2", "Cst3", "Fabp3",
           "Bmpr1b","Fam171b","Ezr","Zic4",
           "Apoe","Rmst", 'Bmp4','Tbx5')



plot_list <- list()
for (i in genes){
  plot_list[[i]] <- Seurat::FetchData(seuratCC, vars = c("UMAP_1","UMAP_2", i)) %>% 
    pivot_longer(-c(UMAP_1,UMAP_2), names_to = 'Gene', values_to = 'Expression') %>% 
    filter(Expression > 0) %>% 
    ggplot(aes(x=UMAP_1,y=UMAP_2)) + 
    geom_point(data = Seurat::FetchData(seuratCC, vars = c("UMAP_1","UMAP_2")),
               color = 'gray90', size = 0.4) +
    geom_point( aes(color = Expression), size = 0.4) +
    scale_color_gradient(low = 'gray85', high = 'navyblue', name = '') +
    #scale_color_brewer(low = 'slateblue1', high = 'slateblue4', name = '') +
    cowplot::theme_cowplot() + 
    facet_wrap(~Gene) + 
    xlab("UMAP 1") + 
    ylab("UMAP 2")
}

cowplot::plot_grid(plotlist = plot_list, ncol = 5)
```


# Heatmap   
Top 10 diff in cluster 5 plus a few selected
```{r}
genes <- seurat.markersCC %>%  
    left_join(seuratCC@meta.data %>% as_tibble() %>% 
                  dplyr::select(cluster = seurat_clusters, CellType) %>% unique()) %>% 
    filter(cluster == 5, p_val_adj < 0.01) %>% head(12) %>% pull(Gene)

genes <- c(genes, c("Shtn1","Fzd8", "Sox1", "Lrrn1")) %>% unique()

library(ComplexHeatmap)
counts <- Seurat::FetchData(seuratCC, vars = c('CellType', genes)) %>% arrange(CellType) 
ct_color <- pals::glasbey(n = length(unique(counts$CellType)))
names(ct_color) <- counts$CellType %>% unique() %>% sort()

col_anno <- HeatmapAnnotation(df = data.frame(CellType = counts$CellType %>% factor()),
                                  col = list(CellType = ct_color))

hm_counts <- counts[,-1] %>% as.matrix() %>% t()
colnames(hm_counts) <- NULL

col_fun <- circlize::colorRamp2(c(0,2), c("#440154FF","#FDE725FF"))
Heatmap(hm_counts, top_annotation = col_anno, cluster_columns = FALSE, col = col_fun, name = ' ' )


tiff(filename = 'heatmap.tiff', width = 3500, height = 2500, res = 600)
print(Heatmap(hm_counts, top_annotation = col_anno, cluster_columns = FALSE, col = col_fun, name = ' ' ))
dev.off()
```



# Top Cluster 5 (Optic Fissure) Diff Genes

For a table?
```{r, fig.width=8}

seurat.markersCC %>%  
  left_join(seuratCC@meta.data %>% as_tibble() %>% 
              dplyr::select(cluster = seurat_clusters, CellType) %>% unique()) %>% 
  filter(cluster == 5, p_val_adj < 0.01) %>% 
  DT::datatable()

```

## GO Enrichment 
```{r}
library(clusterProfiler)
library(enrichplot)


diff_genes <- seurat.markersCC %>%  
  left_join(seuratCC@meta.data %>% as_tibble() %>% 
              dplyr::select(cluster = seurat_clusters, CellType) %>% unique()) %>% 
  filter(cluster == 5, p_val_adj < 0.01)

eg_diff_genes <- bitr(diff_genes$Gene, fromType="SYMBOL", toType="ENTREZID", OrgDb="org.Mm.eg.db")
eg_diff_genes <- diff_genes %>% left_join(., eg_diff_genes, by = c('Gene' = 'SYMBOL'))
# only keep genes that have a mean of >= 10 reads across all sample
expressed <- seuratCC@assays$RNA[seuratCC@assays$RNA %>% rowSums() > 10,] %>% row.names()
eg_universe = bitr(expressed, fromType="SYMBOL", toType="ENTREZID", OrgDb="org.Mm.eg.db")
ego <- enrichGO(gene          = eg_diff_genes$ENTREZID,
                universe      = eg_universe$ENTREZID,
                OrgDb         = org.Mm.eg.db,
                ont           = "all",
                pAdjustMethod = "BH",
                pvalueCutoff  = 0.01,
                qvalueCutoff  = 0.05,
                readable      = TRUE)
```

```{r, fig.width=10, fig.height=14}
p1 <- dotplot(ego, showCategory=30) + ggtitle("Dotplot for GO")
p1
```

### Table
```{r}
ego <- setReadable(ego, OrgDb = org.Mm.eg.db, keyType = 'ENTREZID')
ego@result %>% DT::datatable()
```
### "CNET" Plot
Relationships between related GO terms with shared genes
```{r, fig.width=10, fig.height=8}
geneList <- eg_diff_genes$avg_log2FC
names(geneList) <- eg_diff_genes$Gene
cnet <- cnetplot(ego, foldChange = geneList) + scale_color_viridis_c(name = 'log2(FoldChange)')
cnet
```



# Sesssion Info
```{r}
devtools::session_info()
```
